Update readme with new information...
that says things like clarifying that this is a fork and it's not associated with the original developers or the original project beyond being based on the same code. Also added my contact below the original developer's contact.

Actually, I ended up changing a lot of the readme over to say WeastroShell instead of Abanu, along with a few other things.